+++ hire_hire/api_contributors/__init__.py
@@ -1,0 +1,1 @@
+import api_contributors.schema
