+++ hire_hire/api_duel/__init__.py
@@ -1,0 +1,1 @@
+import api_duel.schema
