+++ hire_hire/api_interview/__init__.py
@@ -1,0 +1,1 @@
+import api_interview.schema
