--- hire_hire/hire_hire/__init__.py
@@ -1,1 +1,0 @@
-import hire_hire.schema
