+++ hire_hire/hire_hire/settings.py
@@ -169,7 +169,7 @@ REST_FRAMEWORK = {
 SIMPLE_JWT = {
     'AUTH_HEADER_TYPES': ('JWT',),
     'ACCESS_TOKEN_LIFETIME': timedelta(days=1),
-    'BLACKLIST_AFTER_ROTATION': False
+    'BLACKLIST_AFTER_ROTATION': False,
 }
 
 CSRF_TRUSTED_ORIGINS = [
