+++ hire_hire/users/__init__.py
@@ -1,0 +1,1 @@
+import users.schema
